requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
ipykernel>=6.0.0
plotly>=5.0.0
joblib>=1.2.0
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.0.0
imbalanced-learn>=0.9.0
mlflow>=1.30.0
scipy>=1.9.0
statsmodels>=0.13.0
